Initial prototype for ragdoll physics
Angular constraints are broken and physics can be a bit weird but the general system is working.

Buoyancy and other effects (like sector thrust) aren't working yet, only the initial thing velocity and animation velocity are considered, along with simple gravity (which has a hack scale because it's too fast?)

New articulated figure type is defined under misc/af, to define the ragdoll figure/skeleton. No tools for this yet so it's manual setting up, but not too complicated.