Improved polyline projection
OG has some singularity at steep angles, turning polylines into invisible planes parallel to the camera.

Not entirely sure what it's doing so just wrote it from scratch.

It's more expensive (since I'm transforming the corners from world to view instead of transforming the center and building it in view space) but it works so whatever.