GPU Lighting experiment (disabled)
Trying to see how viable this is, looks like we'll need to do a ton of refactoring to get the real normals over to the GPU for this to work out

Leaving the code in here for posterity.

Per-pixel works better since we at least have face normals, but unfortunately still looks ass, and sometimes when it doesn't... it almost looks too nice?

Likely to do what I want to achieve (simple lights on particles), we'll need some simple, surface-only dynamic lights. This might be better done using classic deferred since it has the least overhead with GL 3.3.