Disable opaque render buffers for transparent draws
As it was, draws were attempting to ignore opaque buffers like position/normals by forcing the alpha value with should_write_normals, but since sometimes GL_ONE is used, it doesn't always work.

Instead, disable the render buffers when a draw is blending to guarantee it won't write. It's probably more expensive but with even rudimentary state handling it should be ok.