Half baked attempt at rendering some stuff entirely with the GPU
There's way too many problems getting this working.

The depth buffer values from the original passes are complete junk and don't behave well with the HW transformed versions. There's also an irritating flip matrix applied to get it in to the right coordinate space for the shaders. Matching the projection matrix to whatever garbage projection is happening on CPU is also a huge pain.

Colors are broken, textures are broken.

It also somehow breaks adjoins? They disappear at random. Cool.

May very well abandon, all of the issues are infuriating.