Add frustum culling to the clustering
Occluders still get brutally slow when they take up lots of screen estate (ex. the table guy at the very beginning of lvl 1 or if you crouch over a dead guy).

I don't remember the deferred version having this problem. Maybe it's the sheer number of accesses to the cluster TBO, or the VGPR pressure could just be too high and the bandwidth overhead of classic deferred is a better tradeoff on this PC.

I wouldn't expect it, but maybe that hella thin gbuffer I had was enough to make classic a winner. Can't say I know of too many games where classic deferred outperforms tiled/clustering, but here we are lol

All that said, perhaps it would be best to run a depth prepass of all z-write draw calls to reduce some overdraw. Maybe even do occluder passes at 1/2 resolution between the z and color pass and only apply it to opaque surfaces.

I don't really want to build a deferred pipeline, but it might be the most efficient. I'll think about it...