Try to optimize clustering (especially between draw layers)
In first person we change the FOV which invalidates the clusters and causes a complete rebuild, which is EXPENSIVE.

Doubled up the cluster bounds for the draw layers to cache the results.

Unfortunately the cluster density has a major impact on the performance here...

The CPU clustering is very expensive in debug mode, probably performs better in release but I haven't tested yet. Needs more optimizations, perhaps even need some multi-threading here...

Maybe I can throw openmp at it lol

The alternative is to give up on GL 3.3 and move to GL 4.0+ and use compute to do it all on the GPU, but I'd like to keep things compatible.