Add the crosshair stuff to the old bitmap path
It's completely inactive right now but I'm planning to do a refactor.

Currently all menu drawing is thrown together into the final framebuffer in std3D with special handling for the menu, the vbuffer stuff is skipped and there's a lot of hacks to make it work.

This kinda sucks and is making it more difficult to fix the multiplayer preview for RenderDroid 2 than it should be.

Instead the VBuffer should be a texture/framebuffer, not a SDL surface, and all copy operations should be done via hardware via VB-free quad drawing or blit operations. This should hopefully remove redundant code paths and simplify menu rendering (and allow render-to-arbitrary canvas, which can be quite fun for COG down the road), but it may be a big task.